pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iobox"
version = "1.0.0"
description = "Personal workspace context tool — search, retrieve and export email, calendar events and files across Gmail and Microsoft 365"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Tim", email = "tim@goodcollective.com.au"},
]
keywords = ["gmail", "email", "markdown", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Typing :: Typed",
]
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "html2text>=2024.0",
    "typer>=0.9",
    "python-dotenv>=1.0",
    "python-dateutil>=2.8",
    "PyYAML>=6.0",
    "rich>=13.0",
    "tomli-w>=1.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
outlook = ["O365>=2.1.8"]
mcp = ["mcp>=1.2"]
ai = ["anthropic>=0.40"]
semantic = ["sqlite-vec>=0.1", "openai>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.0",
    "ruff>=0.5",
    "mypy>=1.0",
]
docs = [
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.25",
]

[project.urls]
Homepage = "https://github.com/timainge/iobox"
Repository = "https://github.com/timainge/iobox"
Documentation = "https://timainge.github.io/iobox"
Changelog = "https://github.com/timainge/iobox/releases"
Issues = "https://github.com/timainge/iobox/issues"

[project.scripts]
iobox = "iobox.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/iobox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term --cov-report=html"
norecursedirs = [".* venv build dist"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
exclude = ["tests/live/", "tests/fixtures/"]
warn_return_any = true
warn_unused_configs = true

# Legacy modules pre-date strict typing — relax per-module rather than suppress wholesale.
# New code in src/iobox/providers/ is held to full strict mode.
# Note: strict=false in overrides does NOT undo flags set by global strict=true;
# each flag must be explicitly reset.
[[tool.mypy.overrides]]
# space_config uses try/except ImportError for tomllib/tomli compat — the inner
# branch is unreachable on Python 3.11+ so the type: ignore comment appears unused.
module = ["iobox.space_config"]
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = ["iobox.cli", "iobox.mcp_server", "iobox.file_manager", "iobox.email_search", "iobox.auth", "iobox.markdown", "iobox.markdown_converter"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false
disallow_any_generics = false
warn_return_any = false
warn_unused_ignores = false
check_untyped_defs = false
no_implicit_reexport = false

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "mypy",
    "ruff",
    "types-PyYAML",
    "types-python-dateutil",
]
